fixed number

+++ ShaellLang/Number.cs
@@ -27,7 +27,7 @@ namespace ShaellLang
         }
 
         public long ToInteger() => Convert.ToInt64(_numberRepresentation);
-        public long ToFloating() => Convert.ToDouble(_numberRepresentation);
+        public double ToFloating() => Convert.ToDouble(_numberRepresentation);
         public string KeyValue { get => Convert.ToString(_numberRepresentation); }
         public string UniquePrefix { get => "N"; }
     }
